+++ MgDev/Oem/dbxml-2.3.10/xqilla/include/xqilla/functions/FuncFactory.hpp
@@ -34,8 +34,8 @@ public:
   virtual const XMLCh *getName() const = 0;
   virtual const XMLCh *getURI() const = 0;
   virtual const XMLCh *getQName() const = 0;
-  virtual unsigned int getMinArgs() const = 0;
-  virtual unsigned int getMaxArgs() const = 0;
+  virtual size_t getMinArgs() const = 0;
+  virtual size_t getMaxArgs() const = 0;
 };
 
 #endif
