+++ MgDev/Oem/dbxml-2.3.10/xqilla/src/functions/FunctionReverse.cpp
@@ -54,7 +54,7 @@ Sequence FunctionReverse::collapseTreeInternal(DynamicContext* context, int flag
 
   Sequence origSeq = getParamNumber(1, context)->toSequence(context);
   // empty or singleton sequences are identical when reversed
-  unsigned int len=origSeq.getLength();
+  size_t len=origSeq.getLength();
   if(len<2)
     return origSeq;
